backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-cors-headers==4.6.0
python-decouple==3.8
psycopg2-binary==2.9.9
Pillow==10.4.0
celery==5.3.6
redis==5.0.7
django-extensions==3.2.3
django-debug-toolbar==4.4.6
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1
black==24.10.0
flake8==7.1.1
isort==5.13.2
